requirements.txt
fastapi
uvicorn
python-multipart
crewai
langchain-google-genai
google-generativeai
pydub
python-dotenv
